utility/knowledgeBase/pyproject.toml
[project]
name = "knowledgebase"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-search-documents>=11.6.0",
    "azure-storage-blob>=12.28.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
